+++ eating/src/main/java/com/project/eat/shop/ShopRepository.java
@@ -95,8 +95,8 @@ public interface ShopRepository extends JpaRepository<ShopVO, Object> {
 
     @Query(nativeQuery=true,
             value="SELECT m.menu_name, m.menu_price, m.menu_desc, m.menu_pic, m.menu_id FROM Shop s JOIN Menu m ON s.shop_name = m.shop_name WHERE s.shop_id = :shopId")
-        List<Object[]> findShopWithMenu(@Param("shopId") int shopId);
+        List<Object[]> findShopWithMenu(@Param("shopId") Long shopId);
 
-    public ShopVO findByShopId(int shopId);
+    public ShopVO findByShopId(Long shopId);
 
 }
